mcp-bridge/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sheetstorm-mcp-bridge"
version = "0.1.0"
description = "Local MCP bridge for SheetStorm — stdio proxy to the remote SheetStorm API"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "SheetStorm Team"},
]
dependencies = [
    "mcp[cli]>=1.2.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
]

[project.scripts]
sheetstorm-bridge = "sheetstorm_bridge.__main__:main"

[tool.setuptools.packages.find]
include = ["sheetstorm_bridge*"]
